+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * The API setup now has a checkbox to enable / disable the objectives API. Also, all the quick shortcut buttons have now been enhanced to set/unset this checkbox appropriately for each integration.
 * Removed the quick-setup button for the Spectrum integration as the service appears to be dead and the maintainer has vanished.
 * `/scripts` folder is now excluded from the distribution zip.
+* Added options for fleetcarrier cooldown notifications: Overlay, Popup, Both, None.
 
 ### Bug Fixes: